refactor: -util.js suffix 제거
